fix(emul-12): rename CtaGrid to fix capitalization issue
