+++ src/pages/index.tsx
@@ -11,7 +11,7 @@ import { Layout } from '../components/templates/Layout'
 // import Card from '../components/molecules/card/card'
 // import Features from '../components/organisms/features/features'
 // import Signup from '../components/organisms/signup/signup'
-import { CtaGrid } from '../components/organisms/ctaGrid/CtaGrid'
+import { CtaGrid } from '../components/organisms/CtaGrid/CtaGrid'
 import { Play } from '../img/play.inline'
 import { Arrow } from '../img/arrow.inline'
 import { Band } from '../components/molecules/Band/Band'
